Add most remaining missing CSS value descriptions…
Including big update to add all the OpenType feature tags for font-feature-setting.

Mostly the only desc missing now are vendor prefixed properties, and a few 'none', 'auto' and 'url()', or things that are either too long to explain in a short space or are self evident like 'small', 'medium', 'large', etc.