Add Linux/OSX wrapper script
This script wraps the main code executable for Linux and OS X, redirecting
output to a file and running in the background. An exception is made when help
or version argument are used, not suppressing the output as the bootstrap.js is
expected to write to stdout and immediately exit.

Fixes #77