Remove incorrect ANSI escape code for LF
ANSI technically doesn't have an escape code for LF, since it just uses the 0x0A character. The regex \xA is incorrect (it should be \x0A) and is being treated as the two-character sequence "xA". But then once we've fixed that, replacing 0x0A with \n in removeAnsiEscapeCodes is a no-op. Better to just remove the str.replace(LF, '\n') call.

Fixes #12608.