add conditional to rpm spec to test for arch, fixes 13616
libXss.so.1 was recently introduced as a dependency for correct rpm installation of vscode.

libXss.so.1 is a requirement for runtime.

this conditional check will help  tooling such as yum/dnf/rpm etc ensure the proper dependency is in place at installation time and help avoid bringing in i386 dependency libraries for x86_64 architectures.

i386
libXss.so.1

x86_64
libXss.so.1()(64bit)