Use Word Based Completion Range by Default in TS (#17918)
Fixes #17906

**bug**
To support completion items that span multiple words, I added extra logic in 1.9 insiders to use the word prefix to compute the completion range. This introduced a few regressions for completions that do start with the replacement.

**fix**
Use the old, word based range logic by default. Use the prefix completion if it is longer.