Ensure correct language mode is used for onEnter (#26358)
Fixes #26351

**bug**

For the code

```js
const x = <b>|</b>;
```

with the cursor at the `|`, on enter rules should change the code to:

```js
const x = <b>
    |
</b>;
```

However we currently do not handle this case. The cause is the extra semicolon at the end of the line. This causes `getEnterAction` to think we are in a javascript context for the onenter rules instead of in a js-tags context

**fix**
Pass in the column of the cursor when getting the language id to use for the on enter rules instead of using the end of the line. This ensures we are in the `jsx-tags` language in the above case