Fix case where terminal is shutdown by user/vscode
This bug was caused because process.kill was changed to process.send({
event: 'shutdown' });. Using kill caused the process exit code to be
null however this was no longer the case when calling process.exit
inside the process itself. The fix is to better manage process state
within TerminalInstance and track whether it was killed by the user (ie.
VS Code) inside TerminalInstance.

Fixes #31737