For line comment, always use start line of language for line comments that trigger block comments
**Bug**
When a line comment command falls back to a block comment, we currently get the language mode from the cursor. In jsx, when you are inside of a javascript expression, this results in the wrong language mode being selected

**Fix**
Make sure we always grab the language mode from the start of the line. This is consistent with what we do for regular line comments

Fixes #36173