Make a code action request for each diagnostic
Fixes #27392

**Bug**
We currently only make a single code action request for all diagnostics, with a range the covers the span of all diagnostics. However the TS Server API actually expects a code action request for each diagnostic (both the code and range of the code action request has to exactly match that of the diagnostic it aims to fix).

The result is that for overlapping diagnotics where there are multiple fixes, we only  would get code actions for the outermost diagnostic

**Fix**
Make a code action request per diagnostic