Simplify terminal launch logic
This change goes away with the terminal launch delay (since the
dynamic xterm import does that) as well as waiting for the panel
to display. By doing so, this ensures that the TerminalInstance
is available when it's needed for the TerminalPanel to focus it
when clicking on the panel button.

Fixes #40416