Fix new comment position and event leak
Use the current comment position when creating
a new thread, not the one that was
when user started editing the comment,
because the position may change, e.g. if the user
added new lines before the comment.

Keep the editor selection when comments are
updated via documment comment provider
event so user flow is not disturbed
when comments are updated.

Add button styler to local disposables to
fix event leak.