Add experimental port mapping support inside of iframed based webviews
Hooks up a prototype of how we could support some types of port mapping inside of webviews. It uses the same service-worker based approach that we use for loading local resources inside of webviews.

This does not work for loading iframes however as iframe `src` requests go to the service worker of the `src` origin (not the service worker of the parent that contains the iframe). Not sure if this can be worked around. Two alternatives if we can't:

* Add api for getting the mapped port.
* Dynamically rewrite the input html

Both are not ideal as they may require eagerly opening tunnels (we open them lazily at the moment)