editors - introduce IEditorInput.isSaving()
So far, file editors have reported isDirty() === false when auto save was enabled e.g. to avoid showing dirty indicators. With the advent of custom editors, this needs to be cleaned up. The new method isSaving() serves as a hint that an editor can be dirty but in the process of being saved. As such, we can decide to hide the dirty indicator in that case. But in other cases we may want to know if the editor is dirty or not, even if auto saved.