Clarify language when creating new git branch
When using the git extension, the intent is clear when clicking on "+ Create new branch..." and immediately being prompted "Please provide a branch name". But, when clicking on "+ Create a new branch from..." the intent is ambiguous, since the statement "Please provide a branch name" could refer to the branch from which you wish to base the new branch on or the name of the new branch, the latter being what actually happens.

By specifying "Please provide a *new* branch name...", the intent is clear: the prompt is asking for the name you wish to call the new branch.