Always set `setIgnoreMenuShortcuts` while the webview is focused
There seems to be potential race where both the native menu and webview can get a keyboard event. Right now, we count on the webview getting it first but this is not reliable.

Instead, always disable the menu shortcuts when a webview is focused. This seems to work pretty well in my testing but I'm checking it in early this iteration to see if it causes any regressions