Let core command such as  undo/redo be overridden by webviews and notebooks (#98288)
* Add CoreCommandService

For #90110

This change introduce a new service (`ICoreCommandService`) that lets high levels parts of the editor (such as webviews) hook into core commands such as undo and redo that are fired from the editor menus.

## Why is this needed?
To implement undo/redo in custom editors / webviews, we currently register special commands (`editor.action.customEditor.undo` and `editor.action.customEditor.redo`). This is not ideal since it means that users have to update multiple commands if they wish to rebind undo/redo

These custom command also are not invoked when the user goes to the `edit` menu and select `undo` or `redo`. Instead, the edit menu always calls the core `UndoCommand`

We cannot make `UndoCommand` know about custom editors because it lives in `base`

## What this change does?
This change adds a new `ICoreCommandService` that lets higher level parts of the editor override core commands such as undo and redo. We use a similar approach in the `IOpenerService`.

Right now only `undo` and `redo` are overridable. If this approach looks ok, we could also extend it to `copy`, `paste`, `cut`, and `select all`

* Add docs and clean up types

* Rework implementation

Switch from using a command service to having each command state if is overrideable or not.

This hooks up overrides for:

- Undo/redo
- cut/copy/paste

for webviews, custom editors, and notebooks

* Add ProxyCommand so that multiple registered commands can share a single implementation

* Fix compilation & missing file being referenced

* Introduce and adopt MultiCommand for Undo

* Adopt MultiCommand for Redo and SelectAll

* Adopt MultiCommand for Cut, Copy and Paste

Co-authored-by: Alex Dima <alexdima@microsoft.com>