Fixes #114332
`yarn` no longer compiles the smoke test as a postinstall script. This breaks the flow of running `yarn && yarn smoketest` to execute the smoketest.

So change `yarn smoketest` to also do the compiling of the smoke test. But that generates too much extra work on the build machine, which needs to invoke `yarn smoketest` twice, once for Electron and once for web.

To workaround that, `yarn smoketest-no-compile` is added, which can be used by the build machine to avoid needlessly compiling.

While here, the nodejs version is already checked in the `yarn` preinstall script, so remove the check from the smoke test.