testing: clean up actions, add run/debug all, rm duplication
The WorkspaceTestCollectionService duplicated collection logic
unnecessarily. Centralize to the TestService's collection.

This also moves the state updates for discovering tests, and a new
update for adding roots, into the diff, which lets us synchronize those
changes with the changes to test collections.

There's a super crazy bug with the "run all" action that I have
not dove into yet.