anything quick pick - only restore editor view state in certain cases
I often search for a file to then type @ to see the symbols and to navigate in that file. If then I use the mouse to click into the editor, the view state is still reset. The fix is to only reset viewstate if the picker is closed via explicit gesture (ESC key).