terminal: standardize on customPtyImplementation
I initially was doing a larger refactor where the `onInstanceRequestStartExtensionTerminal`
event was removed from the terminalService and instead events would only
be fired in the proxy instance itself, which would be encapsulated better.
However, the task system depends on the global even firing so that the
MainThreadTerminalService can get the terminal for tasks that create
pseudoterminals, so this didn't work out. Kept a more focus refactor
for now.