Fixes #125430: Add a lock around the extension registry.
The problem was that `restartExtensionHost()` was called and immediately afterwards enablement changed events were coming in.

This would lead to the delta extensions queue being handled while the extension host was starting, making the extension host start with newly enabled extensions, which were then attempted to be enabled via the delta extensions queue.

The solution is to make starting and enablement changes handling exclusive via a lock