Add warning about ILocalPtyService
Since it's restricted to the electron-sandbox layer it can only be used when valid so
there's no concern about needing @optional here. The warning is to discourage any
usage outside of the terminal as this service deals with raw events to manage/
manipulate ptys. Without going through the TerminalService this could create processes
but they wouldn't get managed or displayed.

Part of #127878