Re-evaluate webgl terminals every session
I'm not sure why exactly but I've seen several reports of users getting stuck
not on the canvas renderer and switching to gpuAcceleration=on fixes the
problem. This indicates that there's a caching problem so this change goes the
safer route of just re-evaluating whether webgl throws in every single window.
The downside of this is that some environments may go:
1. Set WebGL, it throws
2. Set canvas, it's slow briefly
3. Set DOM

This will only happen once per window at least though.

Fixes #123077