fix(scm): :bug: do not await git repo status, just give the repo straight up to scm
on repo discovery, the git extension would run `git status` and wait for that to finish before handing out the repository to scm. this was problematic since the scm default repo selection depends on repos being discovered in a timely fashion. there's no reason not to just hand out the repo to scm and let `git status` finish afterwards

Closes: #120089
Closes: #113803