Workaround history entry merging on Linux
This is a temporary workaround to fix the majority case where bash on Linux merges
history identical entries which we normally use as a signal of a skipped command.
The fix checks if the exit code is undefined and the command equals the last command,
inheriting the last command's exit code of so. This will not work when the exit code
differs but that's a pretty small problem with a more invasive change and we'll
track that in another issue.

Fixes #143707