Try to have more explicit error when we don't have a server
Fixes #143032

This change removes a call that would automatically restart the TS server when invoking a command. Instead we now return a `NoServer` response

When this happens, the server still should be restarted. This change stops us from logging a large number of errors when we get into this state where the server has crashed too many times and we no longer want to automatically restart it