debug: improve step in target UI (#152131)
* debug: improve step in target UI

This is based on the DAP proposal here, which we should ratify before
merging this: https://github.com/microsoft/debug-adapter-protocol/issues/274

This PR changes the context menu action to try to step into the target
at the cursor position, if there's a step target with a range at that
position. Otherwise, it opens the context menu as it did before.

Additionally, it adds a **Debug: Step Into Target** command which, when
called (e.g. on keybinding) will show step targets for the current
frame and allow the user to choose one.

* chore: update debugProtocol.d.ts to current