Enforce all properties on output matchers
This enforces output matchers will always have an anchor, offset and length as
otherwise the matcher will search every line which could be thousands of lines.
This also runs the free port issue only on a failed command, we can amend the
xterm.js script to exit with error if needed to support this as it should do
that anyway.

Fixes #161758