Highlight label should not create extra span nodes (#164657)
* Highlight label should not create extra empty dom nodes

I noticed that the `HighlightedLabel` class creates extra `span` elements for text ranges. These should not be needed. Using text children directly should be faster for creation and also reduce the number of nodes in the document

I also related the conditional spread with a longer version that uses a simple call to push. This is worth doing since `HighlightedLabel` is so widely used in the editor

* Update tests

* Update smoke test selector

* Update css